requirements.txt
aiogram==3.2.0
replicate
python-dotenv
requests
